fixed beforestate observation

+++ tictactoe_ai/gamerules/over.py
@@ -14,7 +14,7 @@ def check_gameover(board: Int8[Array, "3 3"]) -> OverResult:
     
     def some_winner():
         winner = jnp.argmax(winners)
-        winner = jax.lax.cond(winner == 0, lambda: -1, lambda: 1)
+        winner = jax.lax.cond(winner == 0, lambda: jnp.int8(-1), lambda: jnp.int8(1))
         return {
             'is_over': True,
             'winner': winner,
@@ -24,7 +24,7 @@ def check_gameover(board: Int8[Array, "3 3"]) -> OverResult:
         no_space_left = check_no_space_left(board)
         return {
             'is_over': no_space_left,
-            'winner': 0,
+            'winner': jnp.int8(0),
         }
     
     return jax.lax.cond(is_winner, some_winner, no_winner)
