+++ tictactoe_ai/gamerules/turn.py
@@ -2,6 +2,7 @@ from .types.state import GameState
 from .over import check_gameover
 from .initialize import initalize_game
 import jax
+from jax import numpy as jnp
 from jaxtyping import Scalar, Int8
 
 
@@ -18,7 +19,7 @@ def turn(state: GameState, action: Int8[Scalar, ""]) -> GameState:
     x_pos = action % 3
     y_pos = action // 3
     board = board.at[y_pos, x_pos].set(active_player)
-    active_player = jax.lax.cond(active_player == 1, lambda: -1, lambda: 1)
+    active_player = jax.lax.cond(active_player == jnp.int8(1), lambda: jnp.int8(-1), lambda: jnp.int8(1))
     
     return {
         'board': board,
